+++ packages/lu/src/parser/lubuild/http-request.ts
@@ -4,42 +4,7 @@
  */
 
 const axios = require('axios')
-const url = require('url')
-const httpsProxyAgent = require('https-proxy-agent')
-
-function httpsProxy(config: any) {
-  /* istanbul ignore if */
-  if (config.socketPath != null) return config
-
-  let parsed = url.parse(config.url)
-  const protocol = parsed.protocol
-  /* istanbul ignore if */
-  if (protocol !== 'https:') return config
-
-  var proxyUrl = process.env['HTTPS_PROXY'] || process.env['https_proxy']
-  /* istanbul ignore if */
-  if (!proxyUrl) return config
-
-  parsed = url.parse(proxyUrl)
-  const proxyOptions = {
-  hostname: parsed.hostname,
-  port: parsed.port,
-  auth: undefined
-  }
-
-  if (parsed.auth) {
-  proxyOptions.auth = parsed.auth
-  }
-
-
-  // HTTPS request must use tunnel proxy protocol
-  config.httpsAgent = httpsProxyAgent(proxyOptions)
-
-  // Disable direct proxy protocol in axios http adapter
-  config.proxy = false
-
-  return config
-}
+const httpsProxy = require('../utils/httpsProxy')
 
 axios.interceptors.request.use(httpsProxy)
